Resolve "Merge doesn't delete already merged person from further evaluation"
There was an error, where a person, who was already deleted because they were merged with another, still got evaluated. During this evaluation, the person which got the index of the deleted one got processed, or a segmentation fault happened, when the deleted person was the last one.

This is fixed by setting `mStatus` to not only signal not being tracked, but also, already being merged away.