Only reduce quality of tracked point on replacing detection
We previously used a comparison to quality 100 for checking, if a point is a detection or not in the changed code. This was replaced by `80`, so a color-detection on e.g. a multicolormarker with code, which has quality 90, still gets counted as detection.

Also a helper function `isDetection` was created, so that this knowledge about the quality <-> detection relationship isn't scattered everywhere in the code.