Replace all Q(Double)SpinBoxes with P(Double)SpinBoxes
The custom spinboxes ignore the wheel event, i.e. can not be changed with the mouse wheel.

This MR does not touch sliders. **Sliders can still be changed by scrolling** and spinboxes connected to sliders (i.e. translation of coordinate system) are changed as well, when the slider is changed by scrolling.