Add extra tab for trajectory tests
Allows to display the result of the current plausibility checks, e.g., inside ROI at begin or end, short trajectories, equality to other person or fast velocity variations, in a table. Clicking on the table jumps to the person(s) and frame of the failed check. A color encoded column shows if the issue was already resolved. The color will be automatically adjusted when a person is changed.