fix: extrinsic calibration failure due to "too high error"
Due to a mistake in !267, the error of the extrinsic calibration was calculated using the old calibration (i.e. all zeros on a new project). This leads to a terrible error and thusly to the rejection of the new calibration.

This commit fixes this by using the new extrinsic parameters to calculate the reprojection error. It also adds some simple regression tests for the extrinsic calibration to catch such mistakes earlier in the future.