Automatically open,close and update Gitlab MRs
When opening a Pull request in the GitHub repo, this will automatically
open a MR in the jugit repo. This contains the description, and also the
MR template from the jugit repo.

When this PR is updated, e.g., description changed, status changed, the
corresponding MR is also updated.

When the PR is closed and not merged, the MR gets also closed. When it
is reopenend, also the MR gets reopened.

Overall this is a convenience change, for everyone using GitHub to
contribute.